[update] 리뷰 별점, 지속성 TYPE FLOAT로 변경

+++ models/review.js
@@ -11,10 +11,10 @@ module.exports = class Review extends Sequelize.Model {
 					type: Sequelize.STRING(45),
 				},
 				stars: {
-					type: Sequelize.DECIMAL(2, 1),
+					type: Sequelize.FLOAT,
 				},
 				longevity: {
-					type: Sequelize.INTEGER,
+					type: Sequelize.FLOAT,
 				},
 				category: {
 					type: Sequelize.INTEGER,
